post

+++ app/Http/Controllers/TasksController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;    // 追加
+use Auth;
 
 class TasksController extends Controller
 {
@@ -15,13 +16,25 @@ class TasksController extends Controller
      */
     public function index()
     {
+        if (\Auth::check()) { // 認証済みの場合
         //task一覧を取得
-        $tasks = Task::all();
+        // $tasks = Task::all();
+        
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+        $tasks = $user->tasks()->orderBy('id', 'desc')->paginate(10);
+        // $tasks = task::orderBy('id', 'desc')->paginate(10);
+        
         // task一覧ビューでそれを表示
         return view('tasks.index', [
             'tasks' => $tasks,
             ]);
     }
+    
+    return view('welcome');
+    
+    }
+
 
     /**
      * Show the form for creating a new resource.
@@ -30,12 +43,21 @@ class TasksController extends Controller
      */
     public function create()
     {
+        // // idの値で投稿を検索して取得
+        // $task = Task::findOrFail($id);
+        // // 認証済みユーザ（閲覧者）がその投稿の所有者でない場合はトップページにリダイレクト
+        // if (\Auth::id() !== $task->user_id) {
+        //     return redirect('/');
+        // }
+        
+        // if (\Auth::check()) { // 認証済みの場合
         $task = new Task;
         
         //task作成ビューを表示
         return view('tasks.create',[
             'task' => $task,
         ]);
+        // }return view('welcome');
     }
 
     /**
@@ -46,6 +68,13 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
+        // // idの値で投稿を検索して取得
+        // $taskid = Task::findOrFail($id);
+        // // 認証済みユーザ（閲覧者）がその投稿の所有者でない場合はトップページにリダイレクト
+        // if (\Auth::id() !== $taskid->user_id) {
+        //     return redirect('/');
+        // }      
+        // if (\Auth::check()) { // 認証済みの場合
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',// 追加
@@ -53,12 +82,14 @@ class TasksController extends Controller
         
         //taskを作成
         $task = new Task;
+        $task->user_id = Auth::id();// 追加
         $task->status = $request->status;// 追加
         $task->content = $request->content;
         $task->save();
         
         //トップページへリダイレクトさせる
         return redirect('/');
+    // }return view('welcome');
     }
 
     /**
@@ -69,13 +100,21 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        //idの値でtaskを検索して取得
+        // idの値で投稿を検索して取得
         $task = Task::findOrFail($id);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者でない場合はトップページにリダイレクト
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }     
+        // if (\Auth::check()) { // 認証済みの場合
+        //idの値でtaskを検索して取得
+        // $task = Task::findOrFail($id);
         
         // task詳細ビューでそれを表示
         return view('tasks.show',[
             'task' => $task,
         ]);
+        // }return view('welcome');
     }
 
     /**
@@ -86,13 +125,21 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        //idの値でtaskを検索して取得
+        // idの値で投稿を検索して取得
         $task = Task::findOrFail($id);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者でない場合はトップページにリダイレクト
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }        
+        // if (\Auth::check()) { // 認証済みの場合
+        //idの値でtaskを検索して取得
+        // $task = Task::findOrFail($id);
         
         // task編集ビューでそれを表示
         return view('tasks.edit',[
             'task' => $task,
         ]);
+        // }return view('welcome');
     }
 
     /**
@@ -104,20 +151,30 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // idの値で投稿を検索して取得
+        $task = Task::findOrFail($id);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者でない場合はトップページにリダイレクト
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }     
+        // if (\Auth::check()) { // 認証済みの場合
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',// 追加
         ]);
         
         //idの値でtaskを検索して取得
-        $task = Task::findOrFail($id);
+        // $task = Task::findOrFail($id);
         // taskを更新
+        $task->user_id = Auth::id();// 追加
         $task->status = $request->status;// 追加
         $task->content = $request->content;
         $task->save();
         
         // トップページへリダイレクトさせる
         return redirect('/');
+        
+        // }return view('welcome');
     }
 
     /**
@@ -128,12 +185,21 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
+
         // idの値でtaskを検索して取得
-        $task = Task::findOrFail($id);
-        // taskを削除
-        $task->delete();
+        $task = \App\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }        
+        
+        // // taskを削除
+        // $task->delete();
         
         // トップページへリダイレクトさせる
         return redirect('/');
+
     }
+    
 }
